+++ CaseStudy/Assets/_GAME/Scripts/Play/Collect/Coin.cs
@@ -4,6 +4,6 @@ namespace _GAME.Scripts.Play.Collect
 {
     public class Coin : MonoBehaviour
     {
-        public float coinValue;
+        public int coinValue;
     }
 }
